+++ client/android/div/src/main/java/com/yandex/div/core/dagger/Div2Component.java
@@ -150,6 +150,7 @@ public interface Div2Component {
     HistogramReporter getHistogramReporter();
 
     @NonNull
+    @Deprecated
     DivPlayerFactory getDivVideoFactory();
 
     @NonNull
@@ -159,6 +160,7 @@ public interface Div2Component {
     DivCreationTracker getDivCreationTracker();
 
     @NonNull
+    @Deprecated
     RenderScript getRenderScript();
 
     @NonNull