+++ client/ios/DivKit/DivKitInfo.swift
@@ -1,3 +1,3 @@
 public enum DivKitInfo {
-  public static let version = "31.1.0"
+  public static let version = "31.2.0"
 }