+++ client/ios/DivKit/Extensions/DivTextExtensions.swift
@@ -119,7 +119,8 @@ extension DivText: DivBlockModeling {
       truncationImages: truncationImages,
       additionalTextInsets: additionalTextInsets,
       canSelect: resolveSelectable(expressionResolver),
-      tightenWidth: resolveTightenWidth(expressionResolver)
+      tightenWidth: resolveTightenWidth(expressionResolver),
+      autoEllipsize: resolveAutoEllipsize(expressionResolver) ?? context.flagsInfo.defaultTextAutoEllipsize
     )
   }
 
@@ -151,12 +152,12 @@ extension DivText: DivBlockModeling {
     guard let text else {
       return []
     }
-    return (images ?? [])
-      .filter {
-        let start = $0.resolveStart(context.expressionResolver) ?? 0
-        return start <= CFAttributedStringGetLength(text)
-      }
-      .map { $0.makeImage(context: context) }
+    return (images ?? []).compactMap { 
+      $0.makeImage(
+        context: context, 
+        textLength: CFAttributedStringGetLength(text)
+      ) 
+    }
   }
 
   private func apply(
@@ -311,17 +312,33 @@ extension URLQueryItem {
 
 extension DivText.Image {
   fileprivate func makeImage(
-    context: DivBlockModelingContext
-  ) -> TextBlock.InlineImage {
+    context: DivBlockModelingContext,
+    textLength: Int
+  ) -> TextBlock.InlineImage? {
     let expressionResolver = context.expressionResolver
+    let start = resolveStart(expressionResolver) ?? 0
+    
+    guard start <= textLength else {
+      return nil
+    }
+    
+    let indexingDirection = resolveIndexingDirection(expressionResolver)
+    let location = switch indexingDirection {
+    case .normal: 
+      start
+    case .reversed:
+      textLength - start
+    }
+    
     return TextBlock.InlineImage(
       size: CGSize(
         width: CGFloat(width.resolveValue(expressionResolver) ?? 0),
         height: CGFloat(height.resolveValue(expressionResolver) ?? 0)
       ),
       holder: context.imageHolderFactory.make(resolveUrl(expressionResolver)),
-      location: resolveStart(expressionResolver) ?? 0,
-      tintColor: resolveTintColor(expressionResolver)
+      location: location,
+      tintColor: resolveTintColor(expressionResolver),
+      tintMode: resolveTintMode(expressionResolver).tintMode
     )
   }
 }
@@ -381,7 +398,7 @@ extension DivLineStyle {
 extension BlockShadow {
   fileprivate var typoShadow: Shadow {
     Shadow(
-      offset: CGSize(width: offset.x, height: -offset.y), 
+      offset: CGSize(width: offset.x, height: -offset.y),
       blurRadius: blurRadius,
       color: color.withAlphaComponent(CGFloat(opacity))
     )