+++ client/ios/DivKit/Views/DivBlockProvider.swift
@@ -167,7 +167,6 @@ final class DivBlockProvider {
 
   private func update(divData: DivData?) {
     guard divData !== self.divData else { return }
-    block = noDataBlock
     guard let divData else {
       self.divData = nil
       return
@@ -178,10 +177,9 @@ final class DivBlockProvider {
     self.divData = divData
   }
 
-  private func update(reasons: [DivActionURLHandler.UpdateReason]) {
+  private func update(reasons: [DivCardUpdateReason]) {
     guard var divData else {
-      guard debugParams.isDebugInfoEnabled else { return }
-      block = makeErrorsBlock(dataErrors)
+      block = debugParams.isDebugInfoEnabled ? makeErrorsBlock(dataErrors) : noDataBlock
       return
     }
 
@@ -237,7 +235,7 @@ final class DivBlockProvider {
     }
   }
 
-  private func needUpdateBlock(reasons: [DivActionURLHandler.UpdateReason]) -> Bool {
+  private func needUpdateBlock(reasons: [DivCardUpdateReason]) -> Bool {
     guard !reasons.isEmpty else { return true }
     for reason in reasons {
       let cardId: DivCardID
@@ -373,7 +371,7 @@ final class DivBlockProvider {
 
 private let noDataBlock = EmptyBlock.zeroSized
 
-extension DivActionURLHandler.UpdateReason {
+extension DivCardUpdateReason {
   fileprivate func patch(for divCardId: DivCardID) -> DivPatch? {
     switch self {
     case let .patch(cardId, patch):