+++ client/ios/LayoutKit/LayoutKit/Blocks/TextBlock.swift
@@ -9,12 +9,20 @@ public final class TextBlock: BlockWithTraits {
     public let holder: ImageHolder
     public let location: Int
     public let tintColor: Color?
+    public let tintMode: TintMode
 
-    public init(size: CGSize, holder: ImageHolder, location: Int, tintColor: Color? = nil) {
+    public init(
+      size: CGSize,
+      holder: ImageHolder,
+      location: Int,
+      tintColor: Color? = nil,
+      tintMode: TintMode = .sourceIn
+    ) {
       self.size = size
       self.holder = holder
       self.location = location
       self.tintColor = tintColor
+      self.tintMode = tintMode
     }
   }
 
@@ -32,6 +40,7 @@ public final class TextBlock: BlockWithTraits {
   public let additionalTextInsets: EdgeInsets
   public let truncationToken: NSAttributedString?
   public let truncationImages: [InlineImage]
+  public let autoEllipsize: Bool
 
   let attachments: [TextAttachment]
   let truncationAttachments: [TextAttachment]
@@ -53,7 +62,8 @@ public final class TextBlock: BlockWithTraits {
     truncationImages: [TextBlock.InlineImage] = [],
     additionalTextInsets: EdgeInsets? = nil,
     canSelect: Bool = false,
-    tightenWidth: Bool = false
+    tightenWidth: Bool = false,
+    autoEllipsize: Bool = true
   ) {
     self.widthTrait = widthTrait
     self.heightTrait = heightTrait
@@ -76,6 +86,7 @@ public final class TextBlock: BlockWithTraits {
       self.truncationAttachments = []
     }
     self.additionalTextInsets = additionalTextInsets ?? .zero
+    self.autoEllipsize = autoEllipsize
   }
 
   public convenience init(
@@ -91,7 +102,8 @@ public final class TextBlock: BlockWithTraits {
     truncationImages: [TextBlock.InlineImage] = [],
     additionalTextInsets: EdgeInsets? = nil,
     canSelect: Bool = false,
-    tightenWidth: Bool = false
+    tightenWidth: Bool = false,
+    autoEllipsize: Bool = true
   ) {
     self.init(
       widthTrait: widthTrait,
@@ -107,7 +119,8 @@ public final class TextBlock: BlockWithTraits {
       truncationImages: truncationImages,
       additionalTextInsets: additionalTextInsets,
       canSelect: canSelect,
-      tightenWidth: tightenWidth
+      tightenWidth: tightenWidth,
+      autoEllipsize: autoEllipsize
     )
   }
 
@@ -135,26 +148,15 @@ public final class TextBlock: BlockWithTraits {
   public func intrinsicContentHeight(forWidth width: CGFloat) -> CGFloat {
     switch heightTrait {
     case let .intrinsic(_, minSize, maxSize):
-      if let cached = cachedIntrinsicHeight,
-         cached.width.isApproximatelyEqualTo(width) {
-        return cached.height
-      }
-
-      let height = ceil(
-        text.heightForWidth(
-          width,
-          maxNumberOfLines: maxIntrinsicNumberOfLines,
-          minNumberOfHiddenLines: minNumberOfHiddenLines,
-          truncationToken: truncationToken
-        ) + additionalTextInsets.vertical.sum
+      clamp(
+        calculateTextIntrinsicContentHeight(for: width),
+        min: minSize,
+        max: maxSize
       )
-      let result = clamp(height, min: minSize, max: maxSize)
-      cachedIntrinsicHeight = (width: width, height: result)
-      return result
     case let .fixed(value):
-      return value
+      value
     case .weighted:
-      return 0
+      0
     }
   }
 
@@ -184,6 +186,27 @@ public final class TextBlock: BlockWithTraits {
       && lhs.images == rhs.images
       && lhs.accessibilityElement == rhs.accessibilityElement
       && lhs.tightenWidth == rhs.tightenWidth
+      && lhs.autoEllipsize == rhs.autoEllipsize
+  }
+
+  public func calculateTextIntrinsicContentHeight(
+    for width: CGFloat
+  ) -> CGFloat {
+    if let cached = cachedIntrinsicHeight,
+       cached.width.isApproximatelyEqualTo(width) {
+      return cached.height
+    }
+
+    let height = ceil(
+      text.heightForWidth(
+        width,
+        maxNumberOfLines: maxIntrinsicNumberOfLines,
+        minNumberOfHiddenLines: minNumberOfHiddenLines,
+        truncationToken: truncationToken
+      ) + additionalTextInsets.vertical.sum
+    )
+    cachedIntrinsicHeight = (width: width, height: height)
+    return height
   }
 }
 
@@ -294,6 +317,6 @@ extension TextBlock.InlineImage {
   fileprivate var tintColorImage: Image? {
     let holderImage = holder.image
     guard let tintColor else { return holderImage }
-    return holderImage?.redrawn(withTintColor: tintColor)
+    return holderImage?.redrawn(withTintColor: tintColor, tintMode: tintMode)
   }
 }